Re-implement MockOcrClientActivity by deriving it from BaseOcrClientActivity.
