+++ Cloud.Ocr/Cloud.Ocr.Activities/Activities/MockOcrClientActivity.cs
@@ -13,54 +13,12 @@ namespace Cloud.Ocr.Activities
 {
     [LocalizedDisplayName(nameof(Resources.MockOcrClientActivity_DisplayName))]
     [LocalizedDescription(nameof(Resources.MockOcrClientActivity_Description))]
-    public class MockOcrClientActivity : ContinuableAsyncCodeActivity
+    public class MockOcrClientActivity : BaseOcrClientActivity
     {
-        #region Properties
-
-        /// <summary>
-        /// If set, continue executing the remaining activities even if the current activity has failed.
-        /// </summary>
-        [LocalizedCategory(nameof(Resources.Common_Category))]
-        [LocalizedDisplayName(nameof(Resources.ContinueOnError_DisplayName))]
-        [LocalizedDescription(nameof(Resources.ContinueOnError_Description))]
-        public override InArgument<bool> ContinueOnError { get; set; }
-
-        #endregion
-
-
-        #region Constructors
-
-        public MockOcrClientActivity()
+        protected override IOcrClient GetOcrClient()
         {
+            return new MockOcrClient();
         }
-
-        #endregion
-
-
-        #region Protected Methods
-
-        protected override void CacheMetadata(CodeActivityMetadata metadata)
-        {
-
-            base.CacheMetadata(metadata);
-        }
-
-        protected override async Task<Action<AsyncCodeActivityContext>> ExecuteAsync(AsyncCodeActivityContext context, CancellationToken cancellationToken)
-        {
-            // Inputs
-            IOcrClient ocrClient = new MockOcrClient();
-            var objectContainer = context.GetFromContext<IObjectContainer>(OcrScope.ParentContainerPropertyTag);
-            objectContainer.Add(ocrClient);
-
-            ///////////////////////////
-            // Add execution logic HERE
-            ///////////////////////////
-
-            // Outputs
-            return (ctx) => { };
-        }
-
-        #endregion
     }
 }
